feat(margin): align account contracts
Synchronize all current Margin Account requests, validation, response models, public names, examples, and canonical documentation with Binance's live Account catalog and dated leverage changelog.

+++ Binance.Api/Margin/Responses/BinanceIsolatedMarginAccountLimit.cs
@@ -8,10 +8,10 @@ public record BinanceIsolatedMarginAccountLimit
     /// <summary>
     /// Current enabled accounts
     /// </summary>
-    public int EnabledAccount { get; set; }
+    public long EnabledAccount { get; set; }
 
     /// <summary>
     /// Max accounts
     /// </summary>
-    public int MaxAccount { get; set; }
+    public long MaxAccount { get; set; }
 }
